- wire up "link to study" button for SampleType if study module is available - move SampleType study publish actions from experiment into study module - move publish jsp next to their actions - log SampleType publish and recall audit events in PublishAuditProvider and SampleTimelineAuditProvider
