+++ study/src/org/labkey/study/StudyModule.java
@@ -119,13 +119,13 @@ import org.labkey.api.wiki.WikiRenderingService;
 import org.labkey.api.writer.ContainerUser;
 import org.labkey.study.assay.ExperimentListenerImpl;
 import org.labkey.study.assay.StudyPublishManager;
-import org.labkey.study.assay.query.AssayAuditProvider;
+import org.labkey.study.assay.query.PublishAuditProvider;
 import org.labkey.study.audit.StudyAuditProvider;
 import org.labkey.study.controllers.CohortController;
 import org.labkey.study.controllers.CreateChildStudyAction;
 import org.labkey.study.controllers.DatasetController;
 import org.labkey.study.controllers.ParticipantGroupController;
-import org.labkey.study.controllers.PublishController;
+import org.labkey.study.controllers.publish.PublishController;
 import org.labkey.study.controllers.SharedStudyController;
 import org.labkey.study.controllers.StudyController;
 import org.labkey.study.controllers.StudyDefinitionController;
@@ -381,7 +381,7 @@ public class StudyModule extends SpringModule implements SearchService.DocumentP
         wikiService.registerMacroProvider("study", new StudyMacroProvider());
         registerFolderTypes();
         SecurityManager.addViewFactory(new SecurityController.StudySecurityViewFactory());
-        AuditLogService.get().registerAuditType(new AssayAuditProvider());
+        AuditLogService.get().registerAuditType(new PublishAuditProvider());
         AuditLogService.get().registerAuditType(new DatasetAuditProvider());
         AuditLogService.get().registerAuditType(new StudyAuditProvider());
         AuditLogService.get().registerAuditType(new StudySecurityEscalationAuditProvider());
