+++ study/src/org/labkey/study/controllers/publish/PublishController.java
@@ -1,15 +1,16 @@
-package org.labkey.study.controllers;
+package org.labkey.study.controllers.publish;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.labkey.api.action.ApiSimpleResponse;
 import org.labkey.api.action.MutatingApiAction;
+import org.labkey.api.action.SimpleViewAction;
 import org.labkey.api.action.SpringActionController;
 import org.labkey.api.assay.AssayProvider;
 import org.labkey.api.assay.AssayService;
 import org.labkey.api.assay.AssayUrls;
 import org.labkey.api.assay.actions.AssayHeaderView;
 import org.labkey.api.assay.actions.AssayRunsAction;
-import org.labkey.api.assay.actions.BaseAssayAction;
 import org.labkey.api.assay.actions.ProtocolIdForm;
 import org.labkey.api.audit.AuditLogService;
 import org.labkey.api.data.Container;
@@ -17,7 +18,10 @@ import org.labkey.api.data.ContainerFilter;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.exp.api.ExpProtocol;
 import org.labkey.api.exp.api.ExpRun;
+import org.labkey.api.exp.api.ExpSampleType;
 import org.labkey.api.exp.api.ExperimentService;
+import org.labkey.api.exp.api.ExperimentUrls;
+import org.labkey.api.exp.api.SampleTypeService;
 import org.labkey.api.pipeline.PipeRoot;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.pipeline.PipelineService;
@@ -28,6 +32,7 @@ import org.labkey.api.query.QuerySettings;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.RequiresPermission;
+import org.labkey.api.security.User;
 import org.labkey.api.security.permissions.InsertPermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.util.ConfigurationException;
@@ -36,12 +41,13 @@ import org.labkey.api.util.HelpTopic;
 import org.labkey.api.util.URLHelper;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NavTree;
+import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.VBox;
 import org.labkey.api.view.ViewBackgroundInfo;
 import org.labkey.study.assay.AssayPublishConfirmAction;
-import org.labkey.study.assay.StudyPublishManager;
 import org.labkey.study.assay.AssayPublishStartAction;
-import org.labkey.study.assay.query.AssayAuditProvider;
+import org.labkey.study.assay.StudyPublishManager;
+import org.labkey.study.assay.query.PublishAuditProvider;
 import org.springframework.validation.BindException;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -52,9 +58,13 @@ import java.util.List;
 
 public class PublishController extends SpringActionController
 {
-    private static final ActionResolver _resolver = new DefaultActionResolver(PublishController.class,
-        AssayPublishStartAction.class,
-        AssayPublishConfirmAction.class
+    private static final ActionResolver _resolver = new DefaultActionResolver(
+            PublishController.class,
+            AssayPublishStartAction.class,
+            AssayPublishConfirmAction.class,
+            SampleTypePublishStartAction.class,
+            SampleTypePublishConfirmAction.class
+//            SampleTypePublishHistoryAction.class
     );
 
     public PublishController()
@@ -77,8 +87,31 @@ public class PublishController extends SpringActionController
         }
     }
 
+    /**
+     * Create a QueryView over the publish event audit log filtered by the source.
+     */
+    protected @Nullable QueryView createHistoryView(int sourceRowId, String sourceColumnName, ContainerFilter cf, BindException errors)
+    {
+        UserSchema schema = AuditLogService.getAuditLogSchema(getUser(), getContainer());
+        if (schema == null)
+            return null;
+
+        QuerySettings settings = new QuerySettings(getViewContext(), QueryView.DATAREGIONNAME_DEFAULT);
+
+        SimpleFilter filter = new SimpleFilter();
+        if (sourceRowId != -1)
+            filter.addCondition(FieldKey.fromParts(sourceColumnName), sourceRowId);
+        filter.addCondition(cf.createFilterClause(ExperimentService.get().getSchema(), FieldKey.fromParts(PublishAuditProvider.COLUMN_NAME_CONTAINER)));
+
+        settings.setBaseFilter(filter);
+        settings.setQueryName(PublishAuditProvider.PUBLISH_AUDIT_EVENT);
+        QueryView view = schema.createView(getViewContext(), settings, errors);
+        view.setContainerFilter(cf);
+        return view;
+    }
+
     @RequiresPermission(ReadPermission.class)
-    public class PublishAssayHistoryAction extends BaseAssayAction<PublishHistoryForm>
+    public class PublishAssayHistoryAction extends SimpleViewAction<PublishHistoryForm>
     {
         private ExpProtocol _protocol;
 
@@ -93,21 +126,9 @@ public class PublishController extends SpringActionController
             VBox view = new VBox();
             view.addView(new AssayHeaderView(_protocol, form.getProvider(), false, true, containerFilter));
 
-            UserSchema schema = AuditLogService.getAuditLogSchema(getUser(), getContainer());
+            QueryView historyView = createHistoryView(_protocol.getRowId(), PublishAuditProvider.COLUMN_NAME_PROTOCOL, containerFilter, errors);
+            view.addView(historyView);
 
-            if (schema != null)
-            {
-                QuerySettings settings = new QuerySettings(getViewContext(), QueryView.DATAREGIONNAME_DEFAULT);
-
-                SimpleFilter filter = new SimpleFilter();
-                if (_protocol.getRowId() != -1)
-                    filter.addCondition(FieldKey.fromParts(AssayAuditProvider.COLUMN_NAME_PROTOCOL), _protocol.getRowId());
-                filter.addCondition(containerFilter.createFilterClause(ExperimentService.get().getSchema(), FieldKey.fromParts(AssayAuditProvider.COLUMN_NAME_CONTAINER)));
-
-                settings.setBaseFilter(filter);
-                settings.setQueryName(AssayAuditProvider.ASSAY_PUBLISH_AUDIT_EVENT);
-                view.addView(schema.createView(getViewContext(), settings, errors));
-            }
             setHelpTopic(new HelpTopic("publishHistory"));
             return view;
         }
@@ -117,6 +138,77 @@ public class PublishController extends SpringActionController
         {
             root.addChild("Assay List", urlProvider(AssayUrls.class).getBeginURL(getContainer()));
             root.addChild(_protocol.getName(), new ActionURL(AssayRunsAction.class, getContainer()).addParameter("rowId", _protocol.getRowId()));
+            root.addChild("Link to Study History");
+        }
+    }
+
+    public static class SampleTypeHistoryForm
+    {
+        private int rowId;
+        private String containerFilterName;
+
+        public String getContainerFilterName()
+        {
+            return containerFilterName;
+        }
+
+        public void setContainerFilterName(String containerFilterName)
+        {
+            this.containerFilterName = containerFilterName;
+        }
+
+        public int getRowId()
+        {
+            return rowId;
+        }
+
+        public void setRowId(int rowId)
+        {
+            this.rowId = rowId;
+        }
+
+        /** @throws NotFoundException if we can't resolve the protocol */
+        @NotNull ExpSampleType getSampleType(Container scope, User user)
+        {
+            if (rowId == -1)
+                throw new NotFoundException("rowId required");
+
+            ExpSampleType st = SampleTypeService.get().getSampleType(scope, user, rowId);
+            if (st == null)
+                throw new NotFoundException("SampleType not found: " + rowId);
+
+            return st;
+        }
+    }
+
+    @RequiresPermission(ReadPermission.class)
+    public class PublishSampleTypeHistoryAction extends SimpleViewAction<SampleTypeHistoryForm>
+    {
+        private ExpSampleType _sampleType;
+
+        @Override
+        public ModelAndView getView(SampleTypeHistoryForm form, BindException errors)
+        {
+            ContainerFilter containerFilter = ContainerFilter.current(getContainer());
+            if (form.getContainerFilterName() != null)
+                containerFilter = ContainerFilter.getContainerFilterByName(form.getContainerFilterName(), getContainer(), getUser());
+
+            _sampleType = form.getSampleType(getContainer(), getUser());
+            VBox view = new VBox();
+            // CONSIDER: create a header component for the sample type?
+
+            QueryView historyView = createHistoryView(_sampleType.getRowId(), PublishAuditProvider.COLUMN_NAME_SAMPLE_TYPE_ID, containerFilter, errors);
+            view.addView(historyView);
+
+            setHelpTopic(new HelpTopic("publishHistory"));
+            return view;
+        }
+
+        @Override
+        public void addNavTrail(NavTree root)
+        {
+            root.addChild("Sample Types", ExperimentUrls.get().getShowSampleTypeListURL(getContainer()));
+            root.addChild(_sampleType.getName(), ExperimentUrls.get().getShowSampleTypeURL(_sampleType));
             root.addChild("Link to Study History");
         }
     }
